+++ Models/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "car-registration-service-modules",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "mongoose": "^5.12.13"
+  }
+}
