--- phone-pair-service/Email.js
@@ -1,15 +1,0 @@
-// Mongo Stuff
-var mongoose = require('mongoose');
-
-var EmailSchema = new mongoose.Schema({
-    EmailID :Number,
-    CarID : String,   
-    Email : String
-});
-
-var EmailModel = mongoose.model('Email', EmailSchema);
-
-module.exports = {
-    schema: EmailSchema,
-    model: EmailModel
-}
